# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Sport Monks API for Football apps",
    "description": "Shows sample end points for Sport Monks",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:5000",
  "basePath": "",
...
  "consumes": ["application/json"],
  "produces": ["application/json"],
...
  "paths": {
...
    "/leagues/{includes}": {
      "get": {
        "tags": ["Leagues"],
        "summary": "Get all leagues",
        "parameters": [
          {
            "in": "path",
            "name": "includes",
            "required": false,
            "description": "country,season,seasons"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Failed. Not found."
          },
          "500": {
            "description": "Error connecting (Server Error)"
          }
        }
      }
    },
    "/leagues/{id}/{includes}": {
      "get": {
        "tags": ["Leagues"],
        "summary": "Get league by id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "Id of the league searching for (Fa Cup = 24)"
          },
          {
            "in": "path",
            "name": "includes",
            "required": false,
            "description": "country,season,seasons"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Failed. Not found."
          },
          "500": {
            "description": "Error connecting (Server Error)"
          }
        }
      }
    },
    "/leagues/{name}/{includes}": {
      "get": {
        "tags": ["Leagues"],
        "summary": "Get league by name",
        "parameters": [
          {
            "in": "path",
            "name": "team1id",
            "required": true,
            "description": "13"
          },
          {
            "in": "path",
            "name": "team2id",
            "required": true,
            "description": "19"
          },
          {
            "in": "path",
            "name": "includes",
            "required": false,
            "description": "country,season,seasons"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Failed. Not found."
          },
          "500": {
            "description": "Error connecting (Server Error)"
          }
        }
      }
    },
...
  }
}